+++ database/migrations/2019_10_18_143744_create_room.php
@@ -18,7 +18,7 @@ class CreateRoom extends Migration
             $table->timestamps();
             $table->string('appartmentName');
             $table->double('noOfRooms');
-            $table->tinyInteger('squareMeters');
+            $table->unsignedBigInteger('squareMeters');
             $table->string('Description')->nullable();
             $table->double('rentCost')->nullable();
             $table->double('additionalCost')->nullable();
